requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==4.0
django-bootstrap-form==3.4
django-extra-views==0.14.0
django-filter==21.1
django-user-agents==0.4.0
Pillow==8.4.0
pycodestyle==2.8.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
ua-parser==0.10.0
user-agents==2.2.0
